+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.0.2
 * Fix _class property shouldn't be excluded from being added to definition
+* Remove package name and method prefix from slugged definition name
 
 ### 0.0.1
 * Initial version